Add Functional Programming category
As the functional buzz grows it is very common to see nowadays people who have been wondering what functional programming is all about but can't find the right resources. Swift makes it easy to leverage many of the concepts behind functional programming in a pragmatic way - and there are many libs which aim to help our lives by using functional programming in Swift. So I think it is important to have this category.