Added Synaction to the list
# Project URL
https://github.com/Ge0rges/Synaction

# Description
A library that allows connected devices (bonjour, lan) to execute code blocks almost simultaneously by leveraging the precision-time-protocol. This allows for cool applications such as playing music in sync across devices (the provided example).

# Why it should be included to awesome-ios (mandatory)
This is, as far as I know and have seen, the only library of it's kind (yay new category) for iOS in Objective-C or Swift. It tackles a complex and original efficiently and provides a good base layer for developers to prototype upon. 

# Checklist
  - [x] Only one project/change is in this pull request
  - [] Has unit tests, integration tests or UI tests
  - [x] Addition in chronological order (bottom of category)
  - [x] Supports iOS 9 / tvOS 10 or later
  - [x] Supports Swift 4 or later (Objective-C but swift compatible)
  - [x] Has a commit from less than 2 years ago
  - [x] Has a clear README in English